Refactored and renamed the library it self
