+++ sample/build.gradle
@@ -23,5 +23,5 @@ android {
 dependencies {
     compile 'com.android.support:appcompat-v7:22.1.1'
     compile 'com.jakewharton.timber:timber:3.1.0'
-    compile project(':viewpageindicator')
+    compile project(':viewpageindicator_libary')
 }
