+++ sample/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@null"/>
-    <no.jlagos.viewpageindicator.PageIndicator
+    <no.jlagos.viewpagerindicator.PageIndicator
         android:id="@+id/indicator"
         android:layout_width="wrap_content"
         android:layout_height="36dp"
