+++ viewpageindicator_libary/src/main/AndroidManifest.xml
