+++ viewpageindicator_libary/src/main/res/drawable/activecolor.xml
