+++ viewpageindicator_libary/src/main/res/drawable/changingcolor.xml
