+++ viewpageindicator_libary/src/main/res/values/strings.xml
